changed type from 16 to 32bit to satisfy some broken compilers
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@9377 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ camlibs/ptp2/config.c
@@ -216,7 +216,7 @@ struct submenu {
 	char		*name;
 	uint16_t	propid;
 	uint16_t	vendorid;
-	uint16_t	type;
+	uint32_t	type;	/* for 32bit alignment */
 	get_func	getfunc;
 	put_func	putfunc;
 };
